requirements.txt
pymysql==1.1.0
mysqlclient==2.1.1
Flask==3.0.0
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.1.1
Flask-Seeder==1.2.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-Bcrypt==1.0.1
marshmallow==3.20.2
Faker==23.2.1
cryptography==42.0.0
alembic==1.13.1

python-dotenv==0.19.2

pytest==8.0.0
pytest-bdd==7.0.1
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
selenium==4.4.3
webdriver-manager==4.0.1